% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{bands_psd}
\alias{bands_psd}
\title{Compute spectral power of bands listed in the bands argument.}
\usage{
bands_psd(bands, signal, sRate, normalize = c(0.5, 40),
  method = "pwelch")
}
\arguments{
\item{bands}{A list of bands to compute with lower and upper limits in the form `list(c(0,4),c(4,8))``}

\item{signal}{Numerical vector of the signal.}

\item{sRate}{Signal sample rate in Hertz.}

\item{normalize}{A band to normalize (divide) by. Defaults to `c(0.5,40)`. Can be set up to FALSE for raw results.}

\item{method}{pwelch or psm}
}
\value{
A list of bands powers.
}
\description{
`bands_psd` calculates power spectral densities estimates using Welch's method on bands. Bands are computed from spectrogram bands equal or greater than lower limit and inferior to the upper limit.
}
\examples{
bands_psd(bands = list(c(0,4),c(4,8)), signal = sin(c(1:10000)), sRate = 200)
}
