% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{chambon2018}
\alias{chambon2018}
\title{Deep Learning Architecture for Temporal Sleep Stage Classification
implementation in Keras}
\usage{
chambon2018(channels = 6, samples = 6300, weights = FALSE,
  verbose = TRUE)
}
\arguments{
\item{channels}{Integer. Number of channels.}

\item{samples}{Integer. Number of samples in each epoch.}

\item{samples}{Boolean. Returns or not a weighted model.}
}
\value{
A Keras sequential model.
}
\description{
Deep Learning Architecture for Temporal Sleep Stage
Classification implemantation in Keras for R.
}
\references{
Chambon, S., Galtier, M., Arnal, P., Wainrib, G. and Gramfort,
A. (2018) A Deep Learning Architecture for Temporal Sleep Stage
Classification Using Multivariate and Multimodal Time Series.
IEEE Trans. on Neural Systems and Rehabilitation Engineering 26:(758-769).
}
