% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{choi2018}
\alias{choi2018}
\title{Convolutional neural network for real-time apnea-hypopnea event detection during sleep}
\usage{
choi2018(segment_size = 160)
}
\arguments{
\item{segment_size}{Integer. The size of the segment to predict.}
}
\value{
A Keras sequential model.
}
\description{
Keras implementation of the deep learning architecture described by Choi & Al in "Real-time apnea-hypopnea event detection during sleep by convolutional neural network".
}
\references{
Choi SH, Yoon H, Kim HS, et al. Real-time apnea-hypopnea event detection during sleep by convolutional neural networks. Computers in Biology and Medicine. 2018;100:123-131.
}
