% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epochs.R
\name{epochs}
\alias{epochs}
\title{Split signals into a list of epochs according to an events dataframe or an epoch duration.}
\usage{
epochs(signals, sRates, resample = 100, epoch = 30, startTime = 0)
}
\arguments{
\item{signals}{A list of numeric vectors containing signals.}

\item{sRates}{A vector or list of integer values of the signals sample rates.}

\item{resample}{The sample rate to resample all signals. Defaults to 100.}

\item{epochs}{Epochs reference. Can be an events dataframe or the number of seconds of each epoch. Defaults to 30.}
}
\value{
A list of signal chunks
}
\description{
Split signals into a list of epochs according to an events dataframe or an epoch duration.
}
