% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_hypnogram_30s}
\alias{example_hypnogram_30s}
\title{Example hypnogram scored on 30 seconds.}
\format{Dataframe}
\source{
\url{http://cloud.frenchkpi.com/s/wreGqkitWNnWwnP/download}
}
\usage{
example_hypnogram_30s
}
\description{
Example hypnogram scored on 30 seconds.
}
\keyword{datasets}
