% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{plot_event}
\alias{plot_event}
\title{Highlight a scored event over a signal.}
\usage{
plot_event(signal, sRate, sig_start, event_start, event_end, window = 10)
}
\arguments{
\item{signal}{The signal vector.}

\item{sRate}{Sample rate of the signal.}

\item{sig_start}{Date-Time value of the signal start.}

\item{event_start}{Date-Time value of the event start.}

\item{event_end}{Date-Time value of the event end.}

\item{window}{Number of seconds of signal to plot before, and after.}
}
\value{
A plot of the highlighted event over the signal.
}
\description{
Highlight a scored event over a signal.
}
