% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{plot_hypnodensity}
\alias{plot_hypnodensity}
\title{Plot a hypnodensity graph using `ggplot2`.}
\usage{
plot_hypnodensity(hypnodensity)
}
\arguments{
\item{hypnodensity}{A hypnodensity dataframe as returned by the `score_stages_edf` function.}
}
\value{
A `ggplot2` hypnodensity graph.
}
\description{
Plot a hypnodensity graph using `ggplot2` from the values returned from `score_stages_edf` function.
}
\references{
Stephansen, J.B., Olesen, A.N., Olsen, M., Ambati, A., Leary, E.B., Moore, H.E., Carrillo, O., Lin, L., Han, F., Yan, H. and Sun, Y.L., 2018. Neural network analysis of sleep stages enables efficient diagnosis of narcolepsy. Nature communications, 9(1), p.5229.
}
