% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{psm}
\alias{psm}
\title{sine multitaper psd}
\usage{
psm(x, sRate, length = 0)
}
\arguments{
\item{x}{Signal vector.}

\item{sRate}{Sample rate of the signal.}

\item{length}{periodogram resolution. 0 default to not resize.}
}
\value{
peridodogram plotted or raw
}
\description{
sine multitaper psd from psd r package
}
\examples{
psm(sin(c(1:10000)), 200, 100)
}
