% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{pwelch}
\alias{pwelch}
\title{pwelch psd}
\usage{
pwelch(x, sRate, points = 0, overlap = 0, padding = 0, show = TRUE)
}
\arguments{
\item{x}{todo}

\item{sRate}{todo}

\item{points}{todo}

\item{overlap}{todo}

\item{padding}{todo}

\item{show}{todo}
}
\value{
peridodogram plotted or raw
}
\description{
pwelch psd
}
\examples{
pwelch(sin(c(1:10000)), 200)
}
