% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_events_compumedics}
\alias{read_events_compumedics}
\title{Read a stages export from Compumedics software in .txt format.}
\usage{
read_events_compumedics(
  txt,
  startTime,
  labels = c(AWA = 0, N1 = 1, N2 = 2, N3 = 3, REM = 5)
)
}
\arguments{
\item{txt}{txt file path.}

\item{startTime}{Character string or date object of the hypnogram start.}

\item{labels}{Labels and values as a named list. Defaults to c("AWA" = 0, "N1" = 1, "N2" = 2, "N3" = 3, "REM" = 5).}
}
\value{
A dataframe of stages.
}
\description{
Read a stages export from Compumedics software in .txt format.
}
