% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_events_ndb}
\alias{read_events_ndb}
\title{Read events from a Resmed Noxturnal .ndb file.}
\usage{
read_events_ndb(data_file)
}
\arguments{
\item{data_file}{.ndb file path.}
}
\value{
An events dataframe.
}
\description{
Read events from a Resmed Noxturnal .ndb file.
}
