% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_events_noxturnal}
\alias{read_events_noxturnal}
\title{Read a Noxturnal events file (Unicode CSV format)}
\usage{
read_events_noxturnal(dir)
}
\arguments{
\item{dir}{Noxturnal events file path.}
}
\value{
A dataframe of scored events.
}
\description{
Read a Noxturnal events file (Unicode CSV format)
}
