% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_events_profusion}
\alias{read_events_profusion}
\title{Read a annotation file from Compumedics Profusion software in XML format.}
\usage{
read_events_profusion(xml, startTime = as.POSIXlt("1970-01-01 00:00:00"))
}
\arguments{
\item{xml}{XML file path.}

\item{startTime}{Character string or date object of the hypnogram start.}
}
\value{
A dataframe of stages and events.
}
\description{
Read a annotation file from Compumedics Profusion software in XML format.
}
