% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_stages_compumedics}
\alias{read_stages_compumedics}
\title{Read a stages exports from Compumedics software}
\usage{
read_stages_compumedics(txt, startTime)
}
\arguments{
\item{txt}{txt file path}

\item{startTime}{Character string or date object of the hypnogram start.}
}
\value{
A dataframe of stages.
}
\description{
Read a stages exports from Compumedics software
}
