% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_mice}
\alias{score_mice}
\title{Score mice sleep from European Data Format (EDF) files.}
\usage{
score_mice(edf, model = schwabedal2018(weights = TRUE), verbose = TRUE)
}
\arguments{
\item{edf}{Character. European Data Format (EDF) file path.}

\item{model}{model}

\item{verbose}{Boolean. Display or not status messages.}
}
\value{
A dataframe containing predicted hypnodensity values of the record.
}
\description{
Score mice sleep from European Data Format (EDF) files.
}
