% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{score_stages}
\alias{score_stages}
\title{Score stages from polysomnography signals. Provided for research purposes. Do not use in production.}
\usage{
score_stages(signals, sRates, model_path = tempdir(), verbose = TRUE)
}
\arguments{
\item{signals}{A list of the 7 signals vectors. Must be ordered in the following order: `c("C3-M2","C4-M1","O1-M2","E1-M2","E2-M1","1-2")`.}

\item{sRates}{A vector of the sample rates of the 7 signals passed in the `signals` parameter.}

\item{model_path}{The path of the model file. Model will be downloaded if a directory is passed or if the file passed is different from the latest available model.}

\item{verbose}{Boolean. Display or not status messages.}
}
\value{
A matrix containing hypnodensity values for each epoch and each of the 5 stages: `AWA`, `REM`, `N1`, `N2`, `N3`.
}
\description{
Score stages from polysomnography signals using a pre-trained convolutional neural network (See references for model architecture). Model have been pre-trained on data from the Hotel-Dieu, Paris, recorded with Resmed Nox polysomnography device. The model may not be suitable for data recorded with other kind of device. It is provided for research purposes. Do not use in production. Model needs 6 channels to predict sleep stage over 30 seconds epochs: Electroencephalogram:`C3-M2`,`C4-M1`,`O1-M2`; Electrooculogram:`E1-M2`,`E2-M1`, Electromyogram:`1-2`.
}
\references{
Chambon, S., Galtier, M., Arnal, P., Wainrib, G. and Gramfort, A. (2018) A Deep Learning Architecture for Temporal Sleep Stage Classification Using Multivariate and Multimodal Time Series. IEEE Trans. on Neural Systems and Rehabilitation Engineering 26:(758-769).
}
