% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{smooth_liang2012}
\alias{smooth_liang2012}
\title{Smooth hypnogram according to the 11 rules described by Liang & Al.}
\usage{
smooth_liang2012(hypnogram)
}
\arguments{
\item{hypnogram}{A hypnogram dataframe.}
}
\value{
A smoothed hypnogram dataframe.
}
\description{
Smooth hypnogram according to the 11 rules described by Liang & Al.
}
\references{
Liang, Sheng-Fu, Chin-En Kuo, Yu-Han Hu, and Yu-Shian Cheng. “A Rule-Based Automatic Sleep Staging Method.” Journal of Neuroscience Methods 205, no. 1 (March 2012): 169–76. https://doi.org/10.1016/j.jneumeth.2011.12.022.
}
