% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{train_batches}
\alias{train_batches}
\title{Trains a model from files batches.}
\usage{
train_batches(model, batches, epochs = 10)
}
\arguments{
\item{model}{Keras model.}

\item{batches}{Character vector of batches files.}

\item{epochs}{Integer. Number of epochs to train the model.}
}
\value{
A trained and serialized Keras model.
}
\description{
Trains a model from files batches.
}
