% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{train_chambon2018}
\alias{train_chambon2018}
\title{Trains a deep convolutional network for sleep stages classification.}
\usage{
train_chambon2018(batches, epochs = 10)
}
\arguments{
\item{batches}{Character vector of batches files.}

\item{epochs}{Integer. Number of epochs to train the model.}
}
\value{
A weighted Keras model.
}
\description{
Trains a deep convolutional network described by Chambon & Al. in A Deep Learning Architecture for Temporal Sleep Stage Classification Using Multivariate and Multimodal Time Series (2018) on batches of generated by the `generate_batches` function.
}
\references{
Chambon, S., Galtier, M., Arnal, P., Wainrib, G. and Gramfort, A. (2018) A Deep Learning Architecture for Temporal Sleep Stage Classification Using Multivariate and Multimodal Time Series. IEEE Trans. on Neural Systems and Rehabilitation Engineering 26:(758-769). @param records
}
