% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tst90}
\alias{tst90}
\title{Compute TST90, the percentage of time during sleep with an oxygen saturation below 90.}
\usage{
tst90(spo2_signal, sRate, startTime, hypnogram)
}
\arguments{
\item{spo2_signal}{The SpO2 signal vector.}

\item{sRate}{The SpO2 signal vector sample rate.}

\item{startTime}{The SpO2 signal start time.}

\item{hypnogram}{Events dataframe containing hypnogram.}
}
\description{
Compute TST90, the percentage of time during sleep with an oxygen saturation below 90.
}
