% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{write_batches_mice}
\alias{write_batches_mice}
\title{Write batches from mice records}
\usage{
write_batches_mice(records, events, batches_path = "./",
  batch_size = 128, classes_nb = 3, padding = 2, resample = 400,
  verbose = TRUE)
}
\arguments{
\item{verbose}{Boolean. Display or not status messages.}
}
\description{
Write batches from mice records
}
