% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{write_channel}
\alias{write_channel}
\title{Write a timeserie to disk using Morpheo Data Format (MDF) guidelines.}
\usage{
write_channel(channel, signals, headers, mdfPath, endian = .Platform$endian)
}
\arguments{
\item{channel}{character. Channel name.}

\item{signals}{list. European Data Format (EDF) signals list.}

\item{headers}{list. European Data Format (EDF) file headers.}

\item{mdfPath}{character. Morpheo Data Format (MDF) directory path.}

\item{endian}{character. Endianess. \code{"big"} or \code{"little"}. Defaults to platform endian.}
}
\description{
Write a timeserie to disk using Morpheo Data Format (MDF) guidelines.
}
\references{
P. Bouchequet, D. Jin, G. Solelhac, M. Chennaoui, D. Leger, "Morpheo Data Format (MDF), un nouveau format de données simple, robuste et performant pour stocker et analyser les enregistrements de sommeil", Médecine du Sommeil, vol. 15, n 1, p. 48-49, march 2018.
}
