% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{write_hypnogram_compumedics}
\alias{write_hypnogram_compumedics}
\title{Write a XML file containing scored stages for Compumedics software.}
\usage{
write_hypnogram_compumedics(hypnogram, filename)
}
\arguments{
\item{hypnogram}{A rsleep hypnogram dataframe.}

\item{filename}{character File name to write on disk.}
}
\description{
Write a XML file containing scored stages for Compumedics software.
}
