% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{check_events}
\alias{check_events}
\title{Check events dataframe.}
\usage{
check_events(e)
}
\arguments{
\item{e}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Check events dataframe.
}
