% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ecg_200hz}
\alias{example_ecg_200hz}
\title{Sample electrocardiogram signal}
\format{A vector of 2000 values.}
\source{
\url{http://www.sommeil-vigilance.fr/}
}
\usage{
example_ecg_200hz
}
\description{
10 seconds of ECG from Resmed Nox A1 polysomnograph sampled at 200 Hz expressed in Volts.
}
\keyword{datasets}
