% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{write_batches_mice}
\alias{write_batches_mice}
\title{Write batches from mice records}
\usage{
write_batches_mice(
  records,
  events,
  batches_path = "./",
  batch_size = 128,
  classes_nb = 3,
  padding = 2,
  resample = 400,
  verbose = TRUE
)
}
\arguments{
\item{records}{Character. Records paths}

\item{events}{List of events}

\item{batches_path}{Path to write batches}

\item{batch_size}{size of each batch}

\item{classes_nb}{number of classes}

\item{padding}{consecutive epochs to add}

\item{resample}{resample rate}

\item{verbose}{Boolean. Display or not status messages.}
}
\description{
Write batches from mice records
}
